+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### Version 1.24.0
+
+- Angular projects default is "production". Switch to "development" when attempting to debug.
+- Support for projects that use .nvmrc (will switch node version when running commands)
+
 ### Version 1.23.6
 
 - Remote Logging